Close issue 10: Feed display does not register </center> tag
Running lxml_tidy on html is dicey because html5 != xhtml, and we're
serving text/html so browsers are going to go with the looser html5
rules.  lxml_tidy turns <i></i> into <i/>, which is correct for xml
(and thus xhtml), but gets seen by html5 parsers as just an open tag.

We did a bunch of cleanup, added some tests and most importantly, told
lxml_tidy to give us html instead of xml.  We're no longer returning
xhtml, so we'll have to rename the tidy2xhtml function, not that the
name makes much sense anyways.

This should allow tidy2xhtml to return html5 that is also xhtml.